requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.2
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
psycopg2-binary==2.8.6
djangorestframework==3.12.4